Using subparser to handle subcommands.
All functionality for now is being copied to the subcommand
`turbopelican init`, while running `turbopelican` without an argument is
being deprecated.
