+++ src/turbopelican/commands/__init__.py
@@ -1,0 +1,1 @@
+"""Stores all logic for every command in turbopelican."""
