+++ src/turbopelican/commands/init/__init__.py
@@ -1,0 +1,1 @@
+"""This package contains all logic pertinent to initializing a fresh Pelican site."""
